requirements.txt
PyYAML
mysql-connector-python
numpy
scipy
matplotlib
shutil
